feat: :sparkles: Enhance group and academic program models with new fields and responses
