+++ src/modules/academic_program/models/__init__.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class AcademicProgramResponse(BaseModel):
+    modalityAcademic: str
+
