+++ src/modules/group/routes.py
@@ -2,76 +2,14 @@
 
 from typing import List
 from fastapi import APIRouter
-from pydantic import BaseModel, Field
 from src.modules.group.services import get_all_groups, add_group, update_group_by_id, delete_group_by_id, create_academic_schedule_pesnum, get_groups_by_academic_schedule_id, get_academic_schedule_pensum_by_pensum_id_and_academic_schedule_id, create_classroom_x_group
 from src.modules.mirror_group.services import create_mirror_group
 import random 
 import string
 from fastapi import HTTPException
 from starlette import status
+from src.modules.group.models import GroupRequest, GroupCreationRequest, GroupResponse
 
-class GroupRequest(BaseModel):
-    groupSize: int = Field(gt=0)
-    modality: str = Field(min_length=4, max_length=150)
-    code: int = Field(gt=0)
-    mirrorGroupId: int
-    subjectId: int = Field(gt=0)
-    academicSchedulePensumId: int
-    maxSize: int
-    registeredPlaces: int
-
-
-class AcademicSchedulePensumRequest(BaseModel):
-    pensumId: int = Field(gt=0)
-    academicScheduleId: int = Field(gt=0)
-
-class MirrorGroupRequest(BaseModel):
-    name: str
-
-class GroupCreationRequest(BaseModel):
-    group: GroupRequest
-    mirror: MirrorGroupRequest
-    academic: AcademicSchedulePensumRequest
-
-class MirrorGroupResponse(BaseModel):
-    id: int
-    name: str
-
-class AcademicProgramResponse(BaseModel):
-    modalityAcademic: str
-class PensumResponse(BaseModel):
-    academic_program: AcademicProgramResponse
-
-class SubjectResponse(BaseModel):
-    id: int
-    name: str
-    level: int
-    code: str
-    pensum: PensumResponse
-
-class ClassroomResponse(BaseModel):
-    id: int
-    location: str
-    capacity: int
-
-class ClassroomXGroupResponse(BaseModel):
-    id: int
-    mainSchedule: str
-    # classroom_x_group_classroom_x_group_mainClassroomIdToclassroom: ClassroomResponse
-
-class GroupResponse(BaseModel):
-    id: int
-    groupSize: int
-    modality: str 
-    code: int 
-    mirrorGroupId: int 
-    subjectId: int 
-    academicSchedulePensumId: int
-    mirror_group: MirrorGroupResponse
-    subject: SubjectResponse
-    maxSize: int
-    registeredPlaces: int
-    classroom_x_group: List[ClassroomXGroupResponse]
 
 router = APIRouter(
     tags=["group"],
@@ -99,15 +37,12 @@ async def create_group(request: GroupCreationRequest):
         academic_schedule_pensum = await get_academic_schedule_pensum_by_pensum_id_and_academic_schedule_id(
             pensum_id, academic_schedule_id
         )
-
         if academic_schedule_pensum is None:
             academic_schedule_pensum_data = academicSchedulePensumRequest.model_dump()
             academic_schedule_pensum = await create_academic_schedule_pesnum(academic_schedule_pensum_data)
-
         group_request.mirrorGroupId = mirrorGroup.id
         group_request.academicSchedulePensumId = academic_schedule_pensum.id
-        group_data = group_request.model_dump()
-        group = await add_group(group_data)
+        group = await add_group(group_request)
         classroom_x_group = {
             "mainSchedule": "L10-12|M10-12",
             "mainClassroomId": 1,
