+++ src/modules/pensum/models/__init__.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+from src.modules.academic_program.models import AcademicProgramResponse
+
+
+class PensumResponse(BaseModel):
+    academic_program: AcademicProgramResponse
